+++ CHANGELOG.md
@@ -14,6 +14,7 @@
     (use either shortcuts without parameters or generic options method instead)
 * (breaking) core installers bundle now always installed (for both auto scan and manual modes)
     - GuiceBundle noDefaultInstallers() shortcut method or GuiceyOptions.UseCoreInstallers option directly may be used to disable installation
+* (breaking) configuration info api (GuiceyConfigurationInfo.getData()) changed to use java8 Predicate instead of guava    
 * Add methods to GuiceBundle builder:
     - printAvailableInstallers() - diagnostic reporting configured to show only available installers (to easily spot available features)
     - useWebInstallers() - shortcut for installing WebInstallersBundle